Fix for Doug's " Mapping sounds with collisions" problem.
The removed code is believed to be un-necessary. It is benign or mostly harmless for openAL, but causes a resource leak of an FMOD channel (FMOD refcounts playing or paused channels) because it allocates a channel but we never get around to telling it to play and apply() below repeats the process (and then does play the second-allocated channel).